refactor: restructure app with Next.js conventions and separation of concerns: new components, new hook; should be easier to test
