+++ src/types/advocate.ts
@@ -1,0 +1,11 @@
+export interface Advocate {
+	id: number;
+	firstName: string;
+	lastName: string;
+	city: string;
+	degree: string;
+	specialties: string[];
+	yearsOfExperience: number;
+	phoneNumber: number;
+	createdAt?: string;
+}
